requirements.txt
address==0.1.1
alabaster==0.7.12
amqp==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
binaryornot==0.4.4
black==22.1.0
bleach==4.1.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Contacts==1.0
cookiecutter==1.7.3
coverage==6.3.2
crispy-bootstrap5==0.6
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.4
Django==3.2.12
django-admin-interface==0.18.5
django-advanced-report-builder==0.0.5
django-allauth==0.49.0
django-appconf==1.0.2
django-celery-beat==2.2.1
django-colorfield==0.6.3
django-coverage-plugin==2.0.2
django-crispy-forms==1.14.0
django-daterangefilter==1.0.0
django-dbbackup==3.3.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.7.1
django-model-utils==4.2.0
django-money==2.1.1
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-polymorphic==3.1.0
django-redis==5.2.0
django-report-builder==6.4.2
django-report-scaffold==0.1.4
django-report-tools==0.2.2
django-report-utils==0.3.15
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-summernote==0.8.20.0
django-tables2==2.4.1
django-timezone-field==4.2.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djlint==0.7.6
docopt==0.6.2
docutils==0.17.1
et-xmlfile==1.1.0
executing==0.8.2
factory-boy==3.2.1
Faker==13.2.0
filelock==3.6.0
flake8==4.0.1
flake8-isort==4.1.1
gunicorn==19.9.0
identify==2.4.11
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.1.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
Js2Py==0.71
kombu==5.2.4
lazy-object-proxy==1.7.1
ldap3==2.9.1
livereload==2.6.3
MarkupPy==1.14
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
parso==0.8.3
pathspec==0.9.0
phonenumbers==8.12.45
pickleshare==0.7.5
Pillow==9.0.1
pipwin==0.5.1
platformdirs==2.5.1
pluggy==1.0.0
poyo==0.5.0
pre-commit==2.17.0
progressbar2==4.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
py==1.11.0
py-moneyed==1.2
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pyjsparser==2.7.1
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==7.0.1
pytest-django==4.5.2
pytest-sugar==0.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
python-slugify==6.1.0
python-utils==3.1.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.1.4
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
smartencoding==0.3
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.4.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.2.0
tablib==3.2.0
termcolor==1.1.0
testfixtures==6.18.4
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.9.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
types-pytz==2021.3.5
types-PyYAML==6.0.4
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
vine==5.0.0
virtualenv==20.13.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.0.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
